+++ era_customer_success/models/__init__.py
@@ -18,3 +18,4 @@ from . import cs_success_profile
 from . import cs_support_wallet
 from . import cs_value_review
 from . import cs_adoption_assessment
+from . import cs_voc_insight